fix(p2p): bootstrapping logic fix
