+++ p2p/module.go
@@ -1,28 +1,17 @@
 package p2p
 
 import (
-	"context"
-	"encoding/csv"
-	"fmt"
 	"log"
-	"net/http"
-	"regexp"
-	"strconv"
-	"strings"
 
 	"github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/p2p/providers"
 	"github.com/pokt-network/pocket/p2p/providers/addrbook_provider"
 	persABP "github.com/pokt-network/pocket/p2p/providers/addrbook_provider/persistence"
-	rpcABP "github.com/pokt-network/pocket/p2p/providers/addrbook_provider/rpc"
 	"github.com/pokt-network/pocket/p2p/providers/current_height_provider"
-	rpcCHP "github.com/pokt-network/pocket/p2p/providers/current_height_provider/rpc"
 	"github.com/pokt-network/pocket/p2p/raintree"
 	"github.com/pokt-network/pocket/p2p/stdnetwork"
 	"github.com/pokt-network/pocket/p2p/transport"
 	typesP2P "github.com/pokt-network/pocket/p2p/types"
-	"github.com/pokt-network/pocket/rpc"
-	"github.com/pokt-network/pocket/runtime/defaults"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
@@ -215,97 +204,4 @@ func (m *p2pModule) handleNetworkMessage(networkMsgData []byte) {
 	}
 
 	m.GetBus().PublishEventToBus(&event)
-}
-
-func (m *p2pModule) configureBootstrapNodes() error {
-	p2pCfg := m.GetBus().GetRuntimeMgr().GetConfig().P2P
-	var csvReader *csv.Reader
-	customBootstrapNodesCsv := strings.Trim(p2pCfg.BootstrapNodesCsv, " ")
-	if customBootstrapNodesCsv == "" {
-		csvReader = csv.NewReader(strings.NewReader(defaults.DefaultP2PBootstrapNodesCsv))
-	} else {
-		csvReader = csv.NewReader(strings.NewReader(customBootstrapNodesCsv))
-	}
-	bootStrapNodes, err := csvReader.Read()
-	if err != nil {
-		return fmt.Errorf("error parsing bootstrap nodes: %w", err)
-	}
-	for _, node := range bootStrapNodes {
-		if !isValidHostnamePort(node) {
-			return fmt.Errorf("invalid bootstrap node: %s", node)
-		}
-	}
-	m.bootstrapNodes = bootStrapNodes
-	return nil
-}
-
-func isValidHostnamePort(str string) bool {
-	pattern := regexp.MustCompile(`^(https?)://([a-zA-Z0-9.-]+):(\d{1,5})$`)
-	matches := pattern.FindStringSubmatch(str)
-	if len(matches) != 4 {
-		return false
-	}
-	protocol := matches[1]
-	if protocol != "http" && protocol != "https" {
-		return false
-	}
-	port, err := strconv.Atoi(matches[3])
-	if err != nil || port < 0 || port > 65535 {
-		return false
-	}
-	return true
-}
-
-func isSelfInAddrBook(selfAddr cryptoPocket.Address, addrBook typesP2P.AddrBook) bool {
-	for _, peer := range addrBook {
-		if peer.Address.Equals(selfAddr) {
-			return true
-		}
-	}
-	return false
-}
-
-func (m *p2pModule) bootstrap() error {
-	var addrBook typesP2P.AddrBook
-
-	for _, bootstrapNode := range m.bootstrapNodes {
-		m.logger.Info().Str("endpoint", bootstrapNode).Msg("Attempting to bootstrap from bootstrap node")
-
-		client, err := rpc.NewClientWithResponses(bootstrapNode)
-		if err != nil {
-			continue
-		}
-		healthCheck, err := client.GetV1Health(context.TODO())
-		if err != nil || healthCheck == nil || healthCheck.StatusCode != http.StatusOK {
-			m.logger.Warn().Str("bootstrapNode", bootstrapNode).Msg("Error getting a green health check from bootstrap node")
-			continue
-		}
-
-		addressBookProvider := rpcABP.NewRPCAddrBookProvider(
-			rpcABP.WithP2PConfig(
-				m.GetBus().GetRuntimeMgr().GetConfig().P2P,
-			),
-			rpcABP.WithCustomRPCUrl(bootstrapNode),
-		)
-
-		currentHeightProvider := rpcCHP.NewRPCCurrentHeightProvider(rpcCHP.WithCustomRPCUrl(bootstrapNode))
-
-		addrBook, err = addressBookProvider.GetStakedAddrBookAtHeight(currentHeightProvider.CurrentHeight())
-		if err != nil {
-			m.logger.Warn().Err(err).Str("endpoint", bootstrapNode).Msg("Error getting address book from bootstrap node")
-			continue
-		}
-	}
-
-	if len(addrBook) == 0 {
-		return fmt.Errorf("bootstrap failed")
-	}
-
-	for _, peer := range addrBook {
-		m.logger.Debug().Str("address", peer.Address.String()).Msg("Adding peer to addrBook")
-		if err := m.network.AddPeerToAddrBook(peer); err != nil {
-			return err
-		}
-	}
-	return nil
 }
