+++ p2p/utils_test.go
@@ -45,18 +45,22 @@ func init() {
 	keys = generateKeys(nil, maxNumKeys)
 }
 
+func generateKey(_ *testing.T, seed int) cryptoPocket.PrivateKey {
+	seedBytes := make([]byte, ed25519.PrivateKeySize)
+	binary.LittleEndian.PutUint32(seedBytes, uint32(seed))
+	pk, err := cryptoPocket.NewPrivateKeyFromSeed(seedBytes)
+	if err != nil {
+		panic(err)
+	}
+	return pk
+}
+
 func generateKeys(_ *testing.T, numValidators int) []cryptoPocket.PrivateKey {
 	keys := make([]cryptoPocket.PrivateKey, numValidators)
 
 	for i := range keys {
 		seedInt := genesisConfigSeedStart + i
-		seed := make([]byte, ed25519.PrivateKeySize)
-		binary.LittleEndian.PutUint32(seed, uint32(seedInt))
-		pk, err := cryptoPocket.NewPrivateKeyFromSeed(seed)
-		if err != nil {
-			panic(err)
-		}
-		keys[i] = pk
+		keys[i] = generateKey(nil, seedInt)
 	}
 	sort.Slice(keys, func(i, j int) bool {
 		return keys[i].Address().String() < keys[j].Address().String()
